+++ Schema/adminSchema.js
@@ -1,0 +1,12 @@
+const customReferences = require('../references/customReferences');
+
+const adminSchema = customReferences.mongoose.Schema({
+  name: String,
+  email: String,
+  password: String,
+ 
+});
+
+module.exports= {
+     adminSchema :adminSchema
+}
