+++ controllers/Auth/CustomerAuthController.js
@@ -119,7 +119,7 @@ customReferences.app.post(
 
 customReferences.app.post("/viewAllCustomers", async (request, response) => {
   try {
-    const users = await userModel.find(); // Retrieve all Customers from MongoDB
+    const users = await customerModel.find(); // Retrieve all Customers from MongoDB
     response.json(users);
   } catch (error) {
     response.status(500).json({ error: "Internal server error." });
@@ -131,7 +131,7 @@ customReferences.app.delete("/deleteCustomer/:customerId", async (request, respo
     const { customerId } = request.params;
 
     // Find and delete the customer by ID
-    const deletedCustomer = await userModel.findByIdAndDelete(customerId);
+    const deletedCustomer = await customerModel.findByIdAndDelete(customerId);
 
     if (deletedCustomer) {
       response.json({ success: true, message: "Customer deleted successfully." });
@@ -149,7 +149,7 @@ customReferences.app.put("/toggleUserStatus/:userId", async (request, response) 
     const { userId } = request.params;
 
     // Find the user by ID
-    const user = await userModel.findById(userId);
+    const user = await customerModel.findById(userId);
 
     if (!user) {
       return response.json({ success: false, message: "User not found." });
@@ -169,26 +169,3 @@ customReferences.app.put("/toggleUserStatus/:userId", async (request, response) 
 
 
 
-customReferences.app.post("/adminLogin", formData.none(), async (request, response) => {
-  try {
-    const { email, password } = request.body;
-
-    // Find the user by email
-    const admin = await adminModel.findOne({ email });
-
-    if (admin) {
-      // Compare the provided password with the stored hashed password
-      const isPasswordMatch = await bcrypt.compare(password, admin.password);
-
-      if (isPasswordMatch) {
-        response.json({"match": true, "loggedInUser": admin});
-      } else {
-        response.json({"match": false});
-      }
-    } else {
-      response.json({"match": false});
-    }
-  } catch (error) {
-    response.status(500).json({ error: "Internal server error." });
-  }
-});
