--- model/Admin.js
@@ -1,9 +1,0 @@
-const customReferences = require('../references/customReferences');
-const adminSchema = customReferences.mongoose.Schema({
-    "name":String,
-    "email":String,
-    "password":String,
-   
-   });
-
-   module.exports = customReferences.mongoose.model('admins',adminSchema);
