+++ model/adminModel.js
@@ -1,0 +1,5 @@
+const customReferences = require('../references/customReferences');
+
+const adminSchema=require('../Schema/adminSchema');
+module.exports =customReferences.mongoose.model('admins', adminSchema.adminSchema);
+
